+++ .eslintrc.yml
@@ -4,6 +4,5 @@ parserOptions:
 env:
   es6: true
   browser: true
-  commonjs: true
 
 extends: 'htmlacademy/es6'
