+++ src/components/const.js
@@ -1,0 +1,6 @@
+export const TASK_COUNT = 3;
+
+export const Position = {
+  BEFOREEND: `beforeend`,
+  AFTEREND: `afterend`
+};
