+++ src/components/loadMoreButton.js
@@ -1,0 +1,3 @@
+const createLoadMoreButton = () => `<button class="load-more" type="button">load more</button>`;
+
+export default createLoadMoreButton;
